pyproject.toml
[project]
name = "spyder-v2"
version = "0.1.0"
description = "Graph-Based Personal CRM"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "django-environ>=0.11",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.2",
    "python-frontmatter>=1.1.0",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.27",
    "python-docx>=1.1",
    "pdfplumber>=0.11",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
    "ruff>=0.8",
    "pyright>=1.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
